requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
certifi==2026.4.22
charset-normalizer==3.4.7
dataclasses-json==0.6.7
exceptiongroup==1.3.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.1.0
numpy==1.24.4
orjson==3.10.15
packaging==24.2
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
pypdf==5.9.0
PyYAML==6.0.3
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.49
tenacity==8.5.0
typing-extensions==4.13.2
typing-inspect==0.9.0
urllib3==2.2.3
yarl==1.15.2
python-dotenv